Updated Venomancer's Plague Ward.  Still buggy (see description).
The Plague Ward faces an incorrect direction when idling and attacking.
The Poison Sting DoT does not deal damage after the Plague Ward dies,
due to garbage collection issues.

The Plague Ward interaction with Poison Sting is hardcoded in
modifier_poison_sting_debuff_datadriven_on_destroy().  This can be fixed
once we are able to determine the caster of a modifier.